Simplify apollo client initialization
